requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
Flask-PyMongo==2.3.0
PyJWT==1.7.1
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
excel2json-3 == 0.1.6
